+++ models/staging/edfi_3/stage/stg_ef3__staff_education_organization_assignment_associations.sql
@@ -5,6 +5,16 @@ keyed as (
     select 
         {{ gen_skey('k_staff') }},
         {{ edorg_ref() }},
+        {{ dbt_utils.generate_surrogate_key(
+            [
+                'tenant_code',
+                'api_year',
+                'begin_date',
+                'ed_org_id',
+                'staff_unique_id',
+                'staff_classification'
+            ]
+        ) }} as k_staff_ed_org_assignment,
         api_year as school_year,
         base_staff_ed_org_assign.*
         {{ extract_extension(model_name=this.name, flatten=True) }}
@@ -14,7 +24,7 @@ deduped as (
     {{
         dbt_utils.deduplicate(
             relation='keyed',
-            partition_by= 'tenant_code, api_year, begin_date, ed_org_id, staff_unique_id, staff_classification',
+            partition_by= 'k_staff_ed_org_assignment',
             order_by='api_year desc, last_modified_timestamp desc, pull_timestamp desc'
         )
     }}